Module pallet_recovery::pallet::dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

as_recovered

Send a call through a recovered account.

cancel_recovered

Cancel the ability to use as_recovered for account.

claim_recovery

Allow a successful rescuer to claim their recovered account.

close_recovery

As the controller of a recoverable account, close an active recovery process for your account.

create_recovery

Create a recovery configuration for your account. This makes your account recoverable.

initiate_recovery

Initiate the process for recovering a recoverable account.

remove_recovery

Remove the recovery process for your account. Recovered accounts are still accessible.

set_recovered

Allow ROOT to bypass the recovery process and set an a rescuer account for a lost account directly.

vouch_recovery

Allow a “friend” of a recoverable account to vouch for an active recovery process for that account.