Function pallet_recovery::pallet::dispatchables::vouch_recovery

pub fn vouch_recovery<T: Config>(
    lost: <<T as Config>::Lookup as StaticLookup>::Source,
    rescuer: <<T as Config>::Lookup as StaticLookup>::Source
)

Allow a “friend” of a recoverable account to vouch for an active recovery process for that account.

The dispatch origin for this call must be Signed and must be a “friend” for the recoverable account.

Parameters:

lost: The lost account that you want to recover.

rescuer: The account trying to rescue the lost account that you want to vouch for.

The combination of these two parameters must point to an active recovery process.

NOTE: This function is an automatically generated, doc only, uncallable stub. designed to document the vouch_recovery variant of the pallet Call enum. You should not attempt to call this function directly.