Trait pallet_recovery::pallet::Config

pub trait Config: Config {
    type Event: From<Event<Self>> + IsType<Self::Event>;
    type Call: Parameter + Dispatchable<Origin = Self::Origin, PostInfo = PostDispatchInfo> + GetDispatchInfo;
    type Currency: ReservableCurrency<Self::AccountId>;
    type ConfigDepositBase: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>;
    type FriendDepositFactor: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>;
    type MaxFriends: Get<u16>;
    type RecoveryDeposit: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>;
}

Configuration trait.

Associated Types

type Event: From<Event<Self>> + IsType<Self::Event>

The overarching event type.

type Call: Parameter + Dispatchable<Origin = Self::Origin, PostInfo = PostDispatchInfo> + GetDispatchInfo

The overarching call type.

type Currency: ReservableCurrency<Self::AccountId>

The currency mechanism.

type ConfigDepositBase: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>

The base amount of currency needed to reserve for creating a recovery configuration.

This is held for an additional storage item whose value size is 2 + sizeof(BlockNumber, Balance) bytes.

type FriendDepositFactor: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>

The amount of currency needed per additional user when creating a recovery configuration.

This is held for adding sizeof(AccountId) bytes more into a pre-existing storage value.

type MaxFriends: Get<u16>

The maximum amount of friends allowed in a recovery configuration.

type RecoveryDeposit: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>

The base amount of currency needed to reserve for starting a recovery.

This is primarily held for deterring malicious recovery attempts, and should have a value large enough that a bad actor would choose not to place this deposit. It also acts to fund additional storage item whose value size is sizeof(BlockNumber, Balance + T * AccountId) bytes. Where T is a configurable threshold.

Loading content...